templates/requirements.txt
This file tells Render which Python libraries to install.

Flask>=2.0
PyJWT>=2.0
gunicorn>=20.0